улушчение логики контроллеров - добавление проверки, хозяин ли делает запросы, чтоб изменять данные

+++ app/Http/Controllers/Note/DeleteNoteController.php
@@ -8,7 +8,9 @@ class DeleteNoteController extends BaseController
 {
     public function __invoke(Request $request, $noteId){
         $note = $this->service->getNote($noteId);
-        $note->delete();
+
+        if($request->input("login") == $note->user->login)
+            $note->delete();
 
         return redirect()->route("profile.index");
     }
